Handle non Rack::Response::Body body
Sometimes the body is an Array, and sometimes it is a String. Neither of those support calling `body` so we need to handle them separately. The Rack pattern is to either return a string or an object that yields strings when `each` is called on it.

Instead, we simply make a method to call each on any non-string body and turn it into a string.